created a new page for render a sanitaryregister for companion
