+++ src/pages/SanitaryRegisterCompanion.js
@@ -1,0 +1,16 @@
+import Modal from '../components/Modals/Modal'
+import SanitaryRegisterForm from '../components/SanitaryRegister/SanitaryRegisterForm';
+
+function SanitaryRegisterCompanion() {
+
+  return (
+    <Modal>
+      <SanitaryRegisterForm
+        isUser = {false}
+      >
+      </SanitaryRegisterForm>
+    </Modal>
+  )
+}
+
+export default SanitaryRegisterCompanion
